implement `From<u8> for #enum_name` instead of `classify()`
There is a tiny problem with the custom trait which exposes
`classify()`. The moment you have 2 enums with `ClassifyBytes` in the
same file, it means that the `trait U8Ext` was duplicated resulting in
errors.

To solve this, we can also implement the `From<u8> for #enum_name`
trait. This allows us to just use `b'a'.into()` in a context where you
are comparing it with an `Class::AlphaLower` for example.