declaration properties can only contain dashes and a-z values
In case the property starts with `--`, we will parse it as a CSS
variable which can contain a few more characters like emoji for example.

Initial reason that I included uppercase is because I was thinking about
`Webkit`, but that should be `-webkit` in CSS anyway.

Technically uppercase values are valid in CSS:
```css
BODY {
  BACKGROUND-COLOR: RED;
}
```
Play: https://play.tailwindcss.com/sUsWi0Ampe?file=css

But for the sake of the arbitrary property, let's only allow
lower-kebab-case property names.