Ensure `--` is allowed inside candidates (#16972)
This PR fixes an issue where named utilities that contain double dashes
`--` are not extracted correctly.

Some people use `--` in the middle of the utility to create some form of
namespaced utility.

Given this input:
```js
let x = 'foo--bar'
```

The extracted candidates before this change:
```js
[ "let", "x", "--bar" ]
```

The extracted candidates after this change:
```js
[ "let", "x", "foo--bar", "--bar" ]
```

The reason `--bar` is still extracted in both cases is because of the
CSS variable machine. We could improve its extraction by checking its
boundary characters but that's a different issue.

For now, the important thing is that `foo--bar` was extracted.

# Test plan

1. Added new test
2. Existing tests pass