use custom string consumer implementation
The problem here is that our internal `string_machine` is about strings
in Tailwind utilities. This means that spaces are not allowed, but we
can't use that in this context otherwise a string like this:

```
.join(' ')
        ^  3. start of new string
       ^   2. whitespace is not allowed, stop string
      ^    1. start of string
```

This could result in parts of the code to be considered a string when
it's clearly not.

Instead, we will use a custom string consumer implementation that just
keeps reading until the end of the string.