accept `class` in `<div class="…">`
Now that we re-use the boundary characters, it does mean that the
`class` in `<div class="…"></div>` is also extracted.

This wasn't the case before, however there are languages where we do
need this, e.g.:

In Svelte you can do this:

```svelte
<div class:px-4="bool"></div>
```

We pre-process this, such that it becomes:
```svelte
<div class px-4="bool"></div>
```

But we still need to extract the `px-4` here. Best part is, if this
`px-4` was just a `flex` instead, then it wouldn't have worked (due to
hitting this current code path instead of the one for numbers).

Added an additional test for:
```svelte
<div class:flex="bool"></div>
```