different approach
Let's not care about the concept of a string, instead let's only replace
`.` when it's not inside of a bracket stack _or_ when it's not
surrounded by numbers.

That means that `.flex.items-center` will become ` flex items-center`
but `px-2.5` remains as `px-2.5`.

It _does_ mean that some (top-level) strings will be mutated now, but
that shouldn't be an issue for our extractor to run.