add `node_modules` as a default ignored folder
Some people don't have `node_modules` in the `.gitignore` file so in
some scenario's this will always be scanned.

Another example is if you use the Vercel CLI to deploy your project
because `.gitignore` files are not pushed...

This is technically a breaking change, but now that we have `@source not
'…'` you can include files/folders from `node_modules` even though the
folder is ignored by default.