extract used CSS variables from CSS files
We will track CSS files while traversing the folder structure, but don't
extract any normal candidates from these CSS files. We will also not
include these files into any of the returned globs.

We will just run the CSS extractor on these CSS files, and every time we
find a CSS variable, we will verify whether it was used or not.

For now, "using", just means if it is used inside of `var(…)`.