ignore binary extensions, but not folders
We generate a glob to ignore binary extensions that looks like this:
```
*.{mp4,pages,exe,…}
```

However, if you have a folder that happens to end in `.pages` for
example, then this folder will be ignored in its entirety.

To solve this, we will instead generate the following globs:
```
*.{mp4,pages,exe,…}
!*.{mp4,pages,exe,…}/
```

This way, the `/` indicates that we are dealing with a folder and that
these should _not_ be ignored.