WIP: potential failing test
The implementation we chose right now means that there is a situation
where an explicit ignore in your `.gitignore` file could now be included
due to the fix we implemented.

We can always add an additional `@source not` on top of everything, but
ideally this is not needed at all.